+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## v1.2.0
+* [Add Command]: [Format: Space by center character(left only)]
+    - trigger: Ctrl + Shift + P
+    - trigger: context menu(Right click menu)
+
 ## v1.1.0
 * [Add command trigger]: context menu [Space by center character]